refactor & props in route
move data to index.js
