+++ src/components/Dialogs/Dialogs.jsx
@@ -1,36 +1,24 @@
 import { NavLink } from 'react-router-dom';
+import DialogItem from './DialogItem/DialogItem';
 import s from './Dialogs.module.css';
-
-const DialogItem = (props) => {
-    return(
-        <div className={s.dialog}>
-            <NavLink to={"/dialogs/"+props.id} className={s.dialog__person}>{props.name}</NavLink> 
-        </div>
-    );
-}
-const Message = (props) => {
-    return(
-        <div className={s.message}>{props.message}</div>
-    );
-}
-
+import Message from './Message/Message';
 
 const Dialogs = (props) => {
 
-    let dialogData = [
-        {id: 1, name: "user 1"}, 
-        {id: 2, name: "user 2"},
-        {id: 3, name: "user 3"},
-        {id: 4, name: "user 4"},
-        {id: 5, name: "user 5"}
-    ];
+    // let dialogData = [
+    //     {id: 1, name: "user 1"}, 
+    //     {id: 2, name: "user 2"},
+    //     {id: 3, name: "user 3"},
+    //     {id: 4, name: "user 4"},
+    //     {id: 5, name: "user 5"}
+    // ];
 
     let messageData = [
         {id: 1, message: "Hi"}, 
         {id: 2, message: "wassup maaan"}
     ];
-
-    let dialogElements = dialogData.map(
+    console.log(props.dialogData);
+    let dialogElements = props.dialogData.map(
         dialog => <DialogItem name={dialog.name} id={dialog.id} />
     )
