+++ src/components/Dialogs/Message/Message.jsx
@@ -1,0 +1,10 @@
+import { NavLink } from 'react-router-dom';
+import s from './../Dialogs.module.css';
+
+const Message = (props) => {
+    return(
+        <div className={s.message}>{props.message}</div>
+    );
+}
+
+export default Message;
