+++ src/components/Profile/Profile.jsx
@@ -2,10 +2,11 @@ import MyPosts from './MyPosts/MyPosts';
 import s from './Profile.module.css';
 import ProfileInfo from './ProfileInfo/ProfileInfo';
 
-const Profile = () => {
+const Profile = (props) => {
+    console.log(props.link);
     return (
         <div>
-            <ProfileInfo/>
+            <ProfileInfo link={props.link}/>
             <MyPosts />
         </div>
     );
