+++ src/components/Profile/ProfileInfo/ProfileInfo.jsx
@@ -1,10 +1,11 @@
 import s from './ProfileInfo.module.css';
 
-const ProfileInfo = () => {
+const ProfileInfo = (props) => {
+    console.log(props.link);
     return (
         <div>
             <div>
-                <img src="https://img1.badfon.ru/original/1024x768/9/ee/leyk-luiz-kanada-les-nebo.jpg"></img>
+                <img src={props.link}></img>
             </div>
             <div className={s.profile__description}>
                 AVA
